requirements.txt
flask==3.0.0
flask-jwt-extended==4.6.0
flask-cors==4.0.0
pymongo==4.6.1
bcrypt==4.1.2
pillow>=10.3.0
python-dotenv==1.0.0
gunicorn==21.2.0
werkzeug==3.0.1
flask-limiter==3.5.0
python-dateutil==2.8.2
requests==2.31.0
email-validator==2.1.0

# OAuth & Authentication
authlib==1.3.0
requests-oauthlib==1.3.1

# AI/ML & NLP
google-generativeai==0.3.2
scikit-learn>=1.5.0
numpy>=1.26.0
pandas>=2.2.0
joblib>=1.3.2

# Resume Parsing
PyPDF2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.3

# Web Scraping
beautifulsoup4==4.12.3
selenium==4.17.2
lxml>=5.3.0

# Google APIs
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0

# Job Scheduling
APScheduler==3.10.4

# Job APIs
careerjet-api-client==3.0.1

# Redis for caching & queues
redis==5.0.1
celery==5.3.6
